requirements.txt
asgiref==3.9.2
Django==5.2.6
django-redis==6.0.0
djangorestframework==3.16.1
redis==6.4.0
sqlparse==0.5.3
tzdata==2025.2
pika
psycopg2-binary==2.9.10